- added default.png - fixed scaling bug on recording - added an alert when recording time is too short

+++ AmbientPlayer/APWaveFormView.h
@@ -16,7 +16,7 @@
     float _root;
     NSMutableArray *_table;
     UIBezierPath *_path;
-    int _prevX;
+    float _prevX;
     float _maxValue;
     BOOL _showHandle;
 }
